// openapi/generated/model/base-action.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Red Hat Openshift Smart Events Fleet Manager
 * The api exposed by the fleet manager of the RHOSE service.
 *
 * The version of the OpenAPI document: 0.0.1
 * Contact: openbridge-dev@redhat.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface BaseAction
 */
export interface BaseAction {
    /**
     * 
     * @type {string}
     * @memberof BaseAction
     */
    'type': string;
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof BaseAction
     */
    'parameters': { [key: string]: string; };
}